requirements.txt
black
eth-brownie
pandas
python-dateutil
requests
click